api: [ARR] api order

+++ lib/app/pages/features/list_makanan/list_makanan_controller.dart
@@ -54,7 +54,7 @@ class ListMakananController extends GetxController {
     token = prefs.getString('token');
   }
 
-  String formatPrice(double price) {
+  String formatPrice(int price) {
     var formattedPrice = NumberFormat.currency(locale: 'id_ID', symbol: 'Rp ').format(price);
     return formattedPrice.replaceAll(",00", "");
   }
